Add missing return type declatrations, fix dockblock type declaration inconsistencies
